style: Moved around some parts
