+++ src/Application/Exception/StateNotEmptyException.php
@@ -4,5 +4,4 @@ namespace App\Application\Exception;
 
 class StateNotEmptyException extends \Exception
 {
-
 }
