+++ src/Application/Exception/StateNotFoundException.php
@@ -4,5 +4,4 @@ namespace App\Application\Exception;
 
 class StateNotFoundException extends \Exception
 {
-
 }
