+++ src/Application/Handler/SupplyArea/DeleteSupplyAreaHandler.php
@@ -3,15 +3,8 @@
 namespace App\Application\Handler\SupplyArea;
 
 use App\Application\Contract\HandlerInterface;
-use App\Application\Exception\StateNotEmptyException;
-use App\Domain\Command\DeleteDispatchAreaCommand;
-use App\Domain\Command\DeleteStateCommand;
 use App\Domain\Command\SupplyArea\DeleteSupplyAreaCommand;
-use App\Domain\Event\DispatchAreaDeleted;
-use App\Domain\Event\StateDeleted;
 use App\Domain\Event\SupplyArea\SupplyAreaDeleted;
-use App\Domain\Repository\StateRepositoryInterface;
-use App\Repository\DispatchAreaRepository;
 use App\Repository\SupplyAreaRepository;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
