+++ src/Application/Handler/SupplyArea/SwitchStateSupplyAreaHandler.php
@@ -3,19 +3,8 @@
 namespace App\Application\Handler\SupplyArea;
 
 use App\Application\Contract\HandlerInterface;
-use App\Application\Exception\StateNotEmptyException;
-use App\Domain\Command\DeleteDispatchAreaCommand;
-use App\Domain\Command\DeleteStateCommand;
 use App\Domain\Command\SupplyArea\SwitchStateSupplyAreaCommand;
-use App\Domain\Command\SwitchStateDispatchAreaCommand;
-use App\Domain\Command\UpdateDispatchAreaCommand;
-use App\Domain\Event\DispatchAreaDeleted;
-use App\Domain\Event\DispatchAreaSwitchedState;
 use App\Domain\Event\SupplyArea\SupplyAreaSwitchedState;
-use App\Domain\Event\SupplyAreaUpdated;
-use App\Domain\Event\StateDeleted;
-use App\Domain\Repository\StateRepositoryInterface;
-use App\Repository\DispatchAreaRepository;
 use App\Repository\StateRepository;
 use App\Repository\SupplyAreaRepository;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
