+++ src/Application/Handler/SupplyArea/UpdateSupplyAreaHandler.php
@@ -3,16 +3,8 @@
 namespace App\Application\Handler\SupplyArea;
 
 use App\Application\Contract\HandlerInterface;
-use App\Application\Exception\StateNotEmptyException;
-use App\Domain\Command\DeleteDispatchAreaCommand;
-use App\Domain\Command\DeleteStateCommand;
 use App\Domain\Command\SupplyArea\UpdateSupplyAreaCommand;
-use App\Domain\Command\UpdateDispatchAreaCommand;
-use App\Domain\Event\DispatchAreaDeleted;
-use App\Domain\Event\StateDeleted;
 use App\Domain\Event\SupplyArea\SupplyAreaUpdated;
-use App\Domain\Repository\StateRepositoryInterface;
-use App\Repository\DispatchAreaRepository;
 use App\Repository\SupplyAreaRepository;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
