+++ src/Domain/Adapter/CollectionInterface.php
@@ -6,5 +6,4 @@ use Doctrine\Common\Collections\Collection;
 
 interface CollectionInterface extends Collection
 {
-
 }
