+++ src/Domain/Contracts/StateInterface.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Domain\Contracts;
 
-use App\Domain\Adapter\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 
 interface StateInterface
