+++ src/Domain/Event/DispatchArea/DispatchAreaDeleted.php
@@ -2,11 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Event;
+namespace App\Domain\Event\DispatchArea;
 
 use App\Domain\Contracts\DispatchAreaInterface;
 use App\Domain\Contracts\DomainEventInterface;
-use App\Domain\Contracts\StateInterface;
 use App\Domain\Event\Traits\NamedEventTrait;
 use Symfony\Contracts\EventDispatcher\Event;
