+++ src/Domain/Event/State/StateUpdated.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Event;
+namespace App\Domain\Event\State;
 
 use App\Domain\Contracts\DomainEventInterface;
 use App\Domain\Contracts\StateInterface;
 use App\Domain\Event\Traits\NamedEventTrait;
 use Symfony\Contracts\EventDispatcher\Event;
 
-class StateCreated extends Event implements DomainEventInterface
+class StateUpdated extends Event implements DomainEventInterface
 {
     use NamedEventTrait;
 
-    public const NAME = 'state.created';
+    public const NAME = 'state.updated';
 
     private StateInterface $state;
