+++ src/Domain/Event/SupplyArea/SupplyAreaCreated.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace App\Domain\Event\SupplyArea;
 
-use App\Domain\Contracts\DispatchAreaInterface;
 use App\Domain\Contracts\DomainEventInterface;
-use App\Domain\Contracts\StateInterface;
 use App\Domain\Contracts\SupplyAreaInterface;
 use App\Domain\Event\Traits\NamedEventTrait;
 use Symfony\Contracts\EventDispatcher\Event;
