+++ src/Domain/Event/SupplyArea/SupplyAreaSwitchedState.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace App\Domain\Event\SupplyArea;
 
-use App\Domain\Contracts\DispatchAreaInterface;
 use App\Domain\Contracts\DomainEventInterface;
 use App\Domain\Contracts\StateInterface;
 use App\Domain\Contracts\SupplyAreaInterface;
