+++ src/Entity/User.php
@@ -67,10 +67,12 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
      * @ORM\Column(type="boolean", nullable=true)
      */
     private ?bool $toggleAllocSidebar = null;
+
     public function getId(): ?int
     {
         return $this->id;
     }
+
     /**
      * A visual identifier that represents this user.
      *
@@ -80,16 +82,19 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
     {
         return $this->username;
     }
+
     public function setUsername(string $username): self
     {
         $this->username = $username;
 
         return $this;
     }
+
     public function getUserIdentifier(): string
     {
         return $this->username;
     }
+
     /**
      * @see UserInterface
      */
@@ -101,6 +106,7 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
 
         return array_unique($roles);
     }
+
     /**
      * @param array<string> $roles
      *
@@ -112,6 +118,7 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
 
         return $this;
     }
+
     /**
      * @see UserInterface
      */
@@ -119,16 +126,19 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
     {
         return $this->password;
     }
+
     public function setPassword(string $password): self
     {
         $this->password = $password;
 
         return $this;
     }
+
     public function getPlainPassword(): ?string
     {
         return $this->plainPassword;
     }
+
     public function setPlainPassword(?string $plainPassword): self
     {
         $this->plainPassword = $plainPassword;
@@ -136,20 +146,24 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
 
         return $this;
     }
+
     public function getEmail(): ?string
     {
         return $this->email;
     }
+
     public function setEmail(string $email): self
     {
         $this->email = $email;
 
         return $this;
     }
+
     public function getHospital(): ?Hospital
     {
         return $this->hospital;
     }
+
     public function setHospital(Hospital $hospital): self
     {
         // set the owning side of the relation if necessary
@@ -161,26 +175,31 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
 
         return $this;
     }
+
     public function getIsVerified(): ?bool
     {
         return $this->isVerified;
     }
+
     public function setIsVerified(bool $isVerified): self
     {
         $this->isVerified = $isVerified;
 
         return $this;
     }
+
     public function getIsCredentialsExpired(): ?bool
     {
         return $this->isCredentialsExpired;
     }
+
     public function setIsCredentialsExpired(bool $isCredentialsExpired): self
     {
         $this->isCredentialsExpired = $isCredentialsExpired;
 
         return $this;
     }
+
     /**
      * Returning a salt is only needed, if you are not using a modern
      * hashing algorithm (e.g. bcrypt or sodium) in your security.yaml.
@@ -191,6 +210,7 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
     {
         return null;
     }
+
     /**
      * @see UserInterface
      */
@@ -198,24 +218,29 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
     {
         $this->plainPassword = null;
     }
+
     public function __toString(): string
     {
         return $this->getUsername();
     }
+
     public function isVerified(): bool
     {
         return $this->isVerified;
     }
+
     public function getIsParticipant(): ?bool
     {
         return $this->isParticipant;
     }
+
     public function setIsParticipant(?bool $isParticipant): self
     {
         $this->isParticipant = $isParticipant;
 
         return $this;
     }
+
     public function canImport(): bool
     {
         if ($this->hospital) {
@@ -224,36 +249,43 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface, \String
 
         return false;
     }
+
     public function getAllowsEmail(): ?bool
     {
         return $this->allowsEmail;
     }
+
     public function setAllowsEmail(?bool $allowsEmail): self
     {
         $this->allowsEmail = $allowsEmail;
 
         return $this;
     }
+
     public function getAllowsEmailReminder(): ?bool
     {
         return $this->allowsEmailReminder;
     }
+
     public function setAllowsEmailReminder(?bool $allowsEmailReminder): self
     {
         $this->allowsEmailReminder = $allowsEmailReminder;
 
         return $this;
     }
+
     public function getToggleAllocSidebar(): ?bool
     {
         return $this->toggleAllocSidebar;
     }
+
     public function setToggleAllocSidebar(?bool $toggleAllocSidebar): self
     {
         $this->toggleAllocSidebar = $toggleAllocSidebar;
 
         return $this;
     }
+
     public function switchAllocSidebar(): self
     {
         if ($this->toggleAllocSidebar) {
