+++ tests/Unit/Domain/Entity/ImportTest.php
@@ -2,7 +2,6 @@
 
 namespace App\Tests\Unit\Domain\Entity;
 
-use _PHPStan_c862bb974\Nette\Utils\DateTime;
 use App\Domain\Contracts\HospitalInterface;
 use App\Domain\Contracts\UserInterface;
 use App\Domain\Entity\Import;
